tank man turret hitpoints
I'm close to getting the turrets to take damage and die. For some reason the right turret isn't looking like it's getting hit and collisions are acting weird.

Fixed a destroy bug with bullets single. I'm going to have to change how I destroy everything that's a physics object now...